Refactor into libraries
