+++ lib/Engine/Engine.h
@@ -1,0 +1,21 @@
+#ifndef ENGINE_H
+#define ENGINE_H
+
+#include "Config/EngineConfig.h"
+
+#include "Math/Vector.h"
+#include "Math/Matrix.h"
+#include "Math/Quaternion.h"
+#include "Math/Utilities.h"
+#include "Math/Transform.h"
+
+#include "Mesh/Mesh.h"
+
+#include "Rendering/Color.h"
+#include "Rendering/Buffer.h"
+#include "Rendering/Drawing.h"
+#include "Rendering/Output.h"
+#include "Rendering/Pipeline.h"
+#include "Rendering/Projection.h"
+
+#endif // ENGINE_H
