+++ lib/Engine/Math/Quaternion.h
@@ -1,0 +1,17 @@
+#ifndef QUATERNION_H
+#define QUATERNION_H
+
+#include "../Engine.h"
+
+typedef Vector4 Quaternion;
+
+#define IdentityQuaternion \
+    { 0, 0, 0, 1 }
+
+Quaternion FromEuler(Vector3 eulerDegrees);
+Quaternion MulQuaternion(Quaternion a, Quaternion b);
+
+Vector3 Vector3MulQuaternion(Vector3 vec, Quaternion quat);
+Matrix3x3 GetRotationMatrix(Quaternion quat);
+
+#endif // QUATERNION_H
