+++ lib/Engine/Rendering/Output.h
@@ -1,0 +1,8 @@
+#ifndef OUTPUT_H
+#define OUTPUT_H
+
+#include "../Engine.h"
+
+void Render(ColorBuffer* colorBuffer);
+
+#endif // OUTPUT_h
