--- src/TermEngine/Rendering/Output.c
@@ -1,12 +1,0 @@
-#include <stdio.h>
-#include <assert.h>
-
-#include "Buffer.h"
-#include "Color.h"
-
-void Render(ColorBuffer* colorBuffer) {
-    int numElements = colorBuffer->w * colorBuffer->h;
-    for (int i = 0; i < numElements; i++) {
-        DrawColor(colorBuffer->contents[i]);
-    }
-}
