--- src/TermEngine/Rendering/Output.h
@@ -1,8 +1,0 @@
-#ifndef OUTPUT_H
-#define OUTPUT_H
-
-#include "Buffer.h"
-
-unsigned long Render(ColorBuffer* colorBuffer);
-
-#endif // OUTPUT_h
